Ignore commit refs in conflicts section
When searching for the commit hash of the original change for a backport
commit, sometimes the script will pick a commit hash referenced in the
conflicts section. It's very common for the conflicts section to have
commit hashes, so ignore it when searching for the original commit hash.

Conflicts section is parsed using this regex:
r"^(Conflict|Conflicts):.*?\n\n"

The vast majority of Conflicts sections start with "Conflicts:", so
maybe we only need to check for that.

Depends: https://github.com/Kamalheib/tools/pull/5

Signed-off-by: Dennis Chen <dechen@redhat.com>